apps/web/scripts/requirements-smplx.txt
# Requirements for SMPL-X avatar generation
# Install with: pip install -r requirements-smplx.txt

# Core dependencies
torch>=2.0.0
smplx>=0.1.28
pyrender>=0.1.45
trimesh>=3.23.5
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0

# Additional dependencies for pyrender
PyOpenGL==3.1.0
# PyOpenGL-accelerate>=3.1.6  # May cause conflicts, install separately if needed

# For headless rendering (if running on server)
# osmesa  # Optional, for offscreen rendering without display (install separately if needed)
pyglet>=2.0.0  # Window management for pyrender